+++ source/production/F0.Generators/CodeAnalysis/EnumInfoGenerator.cs
@@ -24,14 +24,59 @@ internal sealed partial class EnumInfoGenerator : ISourceGenerator
 		{
 			IReadOnlyCollection<INamedTypeSymbol> symbols = Get_GetName_Symbols(receiver.InvocationArguments, context.Compilation, context.CancellationToken);
 
-			string source = GenerateSourceCode(context.Compilation, context.ParseOptions, symbols);
+			GeneratorOptions generatorOptions = GetOptions(context);
+			string source = GenerateSourceCode(symbols, context.Compilation.Options, context.ParseOptions, generatorOptions);
 
 			var sourceText = SourceText.From(source, Encodings.Utf8NoBom);
 			context.AddSource(HintName, sourceText);
 		}
 	}
 
-	private static string GenerateSourceCode(Compilation compilation, ParseOptions parseOptions, IReadOnlyCollection<INamedTypeSymbol> symbols)
+	private static GeneratorOptions GetOptions(GeneratorExecutionContext context)
+	{
+		return new GeneratorOptions
+		{
+			ThrowIfConstantNotFound = GetThrowIfConstantNotFound(context),
+		};
+
+		static bool GetThrowIfConstantNotFound(GeneratorExecutionContext context)
+		{
+			bool throwIfConstantNotFound;
+
+			bool? config = null;
+			if (context.AnalyzerConfigOptions.GlobalOptions.TryGetValue("f0gen_enum_throw", out string? configSwitch))
+			{
+				config = configSwitch.Equals("true", StringComparison.OrdinalIgnoreCase)
+					? true
+					: configSwitch.Equals("false", StringComparison.OrdinalIgnoreCase)
+						? false
+						: null;
+			}
+
+			bool? build = null;
+			if (context.AnalyzerConfigOptions.GlobalOptions.TryGetValue("build_property.F0Gen_EnumInfo_ThrowIfConstantNotFound", out string? buildSwitch))
+			{
+				build = buildSwitch.Equals("true", StringComparison.OrdinalIgnoreCase) || buildSwitch.Equals("enable", StringComparison.OrdinalIgnoreCase)
+					? true
+					: buildSwitch.Equals("false", StringComparison.OrdinalIgnoreCase) || buildSwitch.Equals("disable", StringComparison.OrdinalIgnoreCase)
+						? false
+						: null;
+			}
+
+			if ((config is true && build is false) || (config is false && build is true))
+			{
+				throwIfConstantNotFound = false;
+			}
+			else
+			{
+				throwIfConstantNotFound = config is true || build is true;
+			}
+
+			return throwIfConstantNotFound;
+		}
+	}
+
+	private static string GenerateSourceCode(IReadOnlyCollection<INamedTypeSymbol> symbols, CompilationOptions compilationOptions, ParseOptions parseOptions, GeneratorOptions generatorOptions)
 	{
 		using StringWriter writer = new(CultureInfo.InvariantCulture);
 		using IndentedTextWriter source = new(writer, Trivia.Tab);
@@ -73,8 +118,8 @@ internal sealed partial class EnumInfoGenerator : ISourceGenerator
 			source.WriteLineNoTabs();
 		}
 
-		Write_GetName_To(source, languageFeatures);
-		Write_GetName_To(source, symbols, compilation, languageFeatures);
+		Write_GetName_To(source, languageFeatures, generatorOptions);
+		Write_GetName_To(source, symbols, compilationOptions, languageFeatures, generatorOptions);
 
 		source.Indent--;
 		source.WriteLine(Tokens.CloseBrace);
@@ -100,5 +145,10 @@ internal sealed partial class EnumInfoGenerator : ISourceGenerator
 		public bool HasNullPropagatingOperator => Version >= LanguageVersion.CSharp6;
 		public bool HasNamespaceAliasQualifier => Version >= LanguageVersion.CSharp2;
 		public bool HasStaticClasses => Version >= LanguageVersion.CSharp2;
+	}
+
+	private sealed class GeneratorOptions
+	{
+		public bool ThrowIfConstantNotFound { get; init; }
 	}
 }